Small bugfixes

+++ ArduinoLibs/SkaarhojAnalog/SkaarhojAnalog.h
@@ -38,20 +38,20 @@ private:
   bool _debugMode;
 
   uint8_t _joystick_index;
-  int _joystick_tolerance;
-  int _joystick_previousPosition[3];
-  int _joystick_previousValue[3];
-  int _joystick_centerValue[3];
+  int16_t _joystick_tolerance;
+  int16_t _joystick_previousPosition[3];
+  int16_t _joystick_previousValue[3];
+  int16_t _joystick_centerValue[3];
   bool _joystick_buttonStatus;
   bool _joystick_buttonStatusLastUp;
   bool _joystick_buttonStatusLastDown;
 
   uint8_t _uniDirectionalSlider_pinIndex;
-  int _uniDirectionalSlider_sliderTolerance;
-  int _uniDirectionalSlider_sliderLowEndOffset;
-  int _uniDirectionalSlider_sliderHighEndOffset;
-  int _uniDirectionalSlider_previousSliderValue;
-  int _uniDirectionalSlider_previousTransitionPosition;
+  int16_t _uniDirectionalSlider_sliderTolerance;
+  int16_t _uniDirectionalSlider_sliderLowEndOffset;
+  int16_t _uniDirectionalSlider_sliderHighEndOffset;
+  int16_t _uniDirectionalSlider_previousSliderValue;
+  int16_t _uniDirectionalSlider_previousTransitionPosition;
   bool _uniDirectionalSlider_sliderDirectionUp;
   bool _uniDirectionalSlider_disableUnidirectionality;
 
@@ -60,7 +60,7 @@ public:
   void debugMode();
 
   // Joystick functions:
-  void joystick_init(int tolerance, uint8_t i2cAddress, uint8_t index = 0);
+  void joystick_init(int16_t tolerance, uint8_t i2cAddress, uint8_t index = 0);
   bool joystick_hasMoved(uint8_t index);
   int joystick_position(uint8_t index);
   bool joystick_buttonUp();
@@ -69,7 +69,7 @@ public:
   int joystick_AnalogRead(uint8_t index);
 
   // Slider functions:
-  void uniDirectionalSlider_init(int sliderTolerance, int sliderLowEndOffset, int sliderHighEndOffset, uint8_t i2cAddress, uint8_t pinIndex);
+  void uniDirectionalSlider_init(int16_t sliderTolerance, int16_t sliderLowEndOffset, int16_t sliderHighEndOffset, uint8_t i2cAddress, uint8_t pinIndex);
   void uniDirectionalSlider_disableUnidirectionality(bool disable);
   bool uniDirectionalSlider_hasMoved();
   int uniDirectionalSlider_position();
